* 'master' of https://github.com/google/oss-fuzz: (330 commits)
  [languages] Handle language builders in cifuzz and build_and_push_test_images. (#6359)
  [base-builder-new] Fix issues (#6363)
  update Shuduo's email address (#6360)
  Update auto-cc for spirv-tools (#6357)
  tor: Use default focal image (#6356)
  net-snmp: Simplify the build scripts (#6231)
  [base-builder-rust] Make install_rust.sh executable. (#6354)
  [infra][rust] Split out rust projects to their own builder image (#6352)
  qemu: add Mauro to the auto_ccs (#6353)
  [infra][python] Split out python projects to their own builder. (#6341)
  [go][infra] Create base-builder-go and use it to build go projects. (#6335)
  [infra][jvm] Split out JVM projects to their own builder. (#6336)
  [cifuzz] Install docker for focal. (#6339)
  [bls-signatures] Enable Fp2 math (#6351)
  [zxing] Enable AWT headless mode (#6350)
  [ClickHouse] Enable coverage for memory and address sanitizer (#6348)
  apache-httpd: make patches inline. (#6349)
  Done (#6347)
  [bitcoin-core] Fix build (#6344)
  Add new C library for subsequent fuzz testing (#6346)
  ...

+++ infra/cifuzz/fuzz_target.py
@@ -21,8 +21,7 @@ import stat
 import subprocess
 import sys
 
-import docker
-
+import base_runner_utils
 # pylint: disable=wrong-import-position,import-error
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 import utils
@@ -33,7 +32,7 @@ logging.basicConfig(
 
 # Use a fixed seed for determinism. Use len_control=0 since we don't have enough
 # time fuzzing for len_control to make sense (probably).
-LIBFUZZER_OPTIONS = '-seed=1337 -len_control=0'
+LIBFUZZER_OPTIONS = ['-seed=1337', '-len_control=0']
 
 # The number of reproduce attempts for a crash.
 REPRODUCE_ATTEMPTS = 10
@@ -54,25 +53,36 @@ class ReproduceError(Exception):
   """Error for when we can't attempt to reproduce a crash."""
 
 
-class FuzzTarget:
+def get_fuzz_target_corpus_dir(workspace, target_name):
+  """Returns the directory for storing |target_name|'s corpus in |workspace|."""
+  return os.path.join(workspace.corpora, target_name)
+
+
+def get_fuzz_target_pruned_corpus_dir(workspace, target_name):
+  """Returns the directory for storing |target_name|'s puned corpus in
+  |workspace|."""
+  return os.path.join(workspace.pruned_corpora, target_name)
+
+
+class FuzzTarget:  # pylint: disable=too-many-instance-attributes
   """A class to manage a single fuzz target.
 
   Attributes:
     target_name: The name of the fuzz target.
     duration: The length of time in seconds that the target should run.
     target_path: The location of the fuzz target binary.
-    out_dir: The location of where output artifacts are stored.
+    workspace: The workspace for storing things related to fuzzing.
   """
 
   # pylint: disable=too-many-arguments
-  def __init__(self, target_path, duration, out_dir, clusterfuzz_deployment,
+  def __init__(self, target_path, duration, workspace, clusterfuzz_deployment,
                config):
     """Represents a single fuzz target.
 
     Args:
       target_path: The location of the fuzz target binary.
       duration: The length of time  in seconds the target should run.
-      out_dir: The location of where the output from crashes should be stored.
+      workspace: The path used for storing things needed for fuzzing.
       clusterfuzz_deployment: The object representing the ClusterFuzz
           deployment.
       config: The config of this project.
@@ -80,44 +90,62 @@ class FuzzTarget:
     self.target_path = target_path
     self.target_name = os.path.basename(self.target_path)
     self.duration = int(duration)
-    self.out_dir = out_dir
+    self.workspace = workspace
     self.clusterfuzz_deployment = clusterfuzz_deployment
     self.config = config
-    self.latest_corpus_path = None
+    self.latest_corpus_path = get_fuzz_target_corpus_dir(
+        self.workspace, self.target_name)
+    os.makedirs(self.latest_corpus_path, exist_ok=True)
+    self.pruned_corpus_path = get_fuzz_target_pruned_corpus_dir(
+        self.workspace, self.target_name)
+    os.makedirs(self.pruned_corpus_path, exist_ok=True)
 
-  def fuzz(self):
+  def _download_corpus(self):
+    """Downloads the corpus for the target from ClusterFuzz and returns the path
+    to the corpus. An empty directory is provided if the corpus can't be
+    downloaded or is empty."""
+    self.clusterfuzz_deployment.download_corpus(self.target_name,
+                                                self.latest_corpus_path)
+    return self.latest_corpus_path
+
+  def prune(self):
+    """Prunes the corpus and returns the result."""
+    self._download_corpus()
+    prune_options = [
+        '-merge=1', self.pruned_corpus_path, self.latest_corpus_path
+    ]
+    result = self.fuzz(use_corpus=False, extra_libfuzzer_options=prune_options)
+    return FuzzResult(result.testcase, result.stacktrace,
+                      self.pruned_corpus_path)
+
+  def fuzz(self, use_corpus=True, extra_libfuzzer_options=None):
     """Starts the fuzz target run for the length of time specified by duration.
 
     Returns:
       FuzzResult namedtuple with stacktrace and testcase if applicable.
     """
     logging.info('Running fuzzer: %s.', self.target_name)
-    command, _ = docker.get_base_docker_run_command(self.out_dir,
-                                                    self.config.sanitizer,
-                                                    self.config.language)
+    if extra_libfuzzer_options is None:
+      extra_libfuzzer_options = []
+    env = base_runner_utils.get_env(self.config, self.workspace)
+    # TODO(metzman): Is this needed?
+    env['RUN_FUZZER_MODE'] = 'interactive'
 
-    # TODO(metzman): Stop using /out for artifacts and corpus. Use another
-    # directory.
-    # If corpus can be downloaded use it for fuzzing.
-    self.latest_corpus_path = self.clusterfuzz_deployment.download_corpus(
-        self.target_name, self.out_dir)
-    if self.latest_corpus_path:
-      command += docker.get_args_mapping_host_path_to_container(
-          self.latest_corpus_path)
-      command += ['-e', 'CORPUS_DIR=' + self.latest_corpus_path]
+    if use_corpus:
+      # If corpus can be downloaded, use it for fuzzing.
+      self._download_corpus()
+      env['CORPUS_DIR'] = self.latest_corpus_path
 
-    command += [
-        '-e', 'RUN_FUZZER_MODE=interactive', docker.BASE_RUNNER_TAG, 'bash',
-        '-c'
-    ]
+    options = LIBFUZZER_OPTIONS.copy() + [
+        f'-max_total_time={self.duration}',
+        # Make sure libFuzzer artifact files don't pollute $OUT.
+        f'-artifact_prefix={self.workspace.artifacts}/'
+    ] + extra_libfuzzer_options
+    command = ['run_fuzzer', self.target_name] + options
 
-    options = LIBFUZZER_OPTIONS + ' -max_total_time=' + str(self.duration)
-    run_fuzzer_command = f'run_fuzzer {self.target_name} {options}'
-
-    command.append(run_fuzzer_command)
-
-    logging.info('Running command: %s', ' '.join(command))
+    logging.info('Running command: %s', command)
     process = subprocess.Popen(command,
+                               env=env,
                                stdout=subprocess.PIPE,
                                stderr=subprocess.PIPE)
 
@@ -135,7 +163,7 @@ class FuzzTarget:
 
     # Crash was discovered.
     logging.info('Fuzzer %s, ended before timeout.', self.target_name)
-    testcase = self.get_testcase(stderr)
+    testcase = get_testcase(stderr)
     if not testcase:
       logging.error(b'No testcase found in stacktrace: %s.', stderr)
       return FuzzResult(None, None, self.latest_corpus_path)
@@ -149,25 +177,28 @@ class FuzzTarget:
     # We found a bug but we won't report it.
     return FuzzResult(None, None, self.latest_corpus_path)
 
-  def free_disk_if_needed(self):
+  def free_disk_if_needed(self, delete_fuzz_target=True):
     """Deletes things that are no longer needed from fuzzing this fuzz target to
     save disk space if needed."""
     if not self.config.low_disk_space:
+      logging.info('Not freeing disk space after running fuzz target.')
       return
-    logging.info(
-        'Deleting corpus, seed corpus and fuzz target of %s to save disk.',
-        self.target_name)
+    logging.info('Deleting corpus and seed corpus of %s to save disk.',
+                 self.target_name)
 
     # Delete the seed corpus, corpus, and fuzz target.
-    if self.latest_corpus_path and os.path.exists(self.latest_corpus_path):
+    for corpus_path in [self.latest_corpus_path, self.pruned_corpus_path]:
       # Use ignore_errors=True to fix
       # https://github.com/google/oss-fuzz/issues/5383.
-      shutil.rmtree(self.latest_corpus_path, ignore_errors=True)
+      shutil.rmtree(corpus_path, ignore_errors=True)
 
-    os.remove(self.target_path)
     target_seed_corpus_path = self.target_path + '_seed_corpus.zip'
     if os.path.exists(target_seed_corpus_path):
       os.remove(target_seed_corpus_path)
+
+    if delete_fuzz_target:
+      logging.info('Deleting fuzz target: %s.', self.target_name)
+      os.remove(self.target_path)
     logging.info('Done deleting.')
 
   def is_reproducible(self, testcase, target_path):
@@ -184,26 +215,20 @@ class FuzzTarget:
       Raises:
         ReproduceError if we can't attempt to reproduce the crash.
     """
-
     if not os.path.exists(target_path):
+      logging.info('Target: %s does not exist.', target_path)
       raise ReproduceError(f'Target {target_path} not found.')
 
     os.chmod(target_path, stat.S_IRWXO)
 
-    command, container = docker.get_base_docker_run_command(
-        self.out_dir, self.config.sanitizer, self.config.language)
-    if container:
-      command += ['-e', f'TESTCASE={testcase}']
-    else:
-      command += ['-v', f'{testcase}:/testcase']
-
-    command += [
-        '-t', docker.BASE_RUNNER_TAG, 'reproduce', self.target_name, '-runs=100'
-    ]
+    env = base_runner_utils.get_env(self.config, self.workspace)
+    env['TESTCASE'] = testcase
+    command = ['reproduce', self.target_name, '-runs=100']
 
     logging.info('Running reproduce command: %s.', ' '.join(command))
     for _ in range(REPRODUCE_ATTEMPTS):
-      _, _, returncode = utils.execute(command)
+      _, _, returncode = utils.execute(command, env=env)
+
       if returncode != 0:
         logging.info('Reproduce command returned: %s. Reproducible on %s.',
                      returncode, target_path)
@@ -241,10 +266,8 @@ class FuzzTarget:
       raise error
 
     if not reproducible_on_code_change:
-      # TODO(metzman): Allow users to specify if unreproducible crashes should
-      # be reported.
       logging.info('Crash is not reproducible.')
-      return False
+      return self.config.report_unreproducible_crashes
 
     logging.info('Crash is reproducible.')
     return self.is_crash_novel(testcase)
@@ -254,8 +277,7 @@ class FuzzTarget:
     can't be reproduced on an older ClusterFuzz build of the target."""
     if not os.path.exists(testcase):
       raise ReproduceError('Testcase %s not found.' % testcase)
-    clusterfuzz_build_dir = self.clusterfuzz_deployment.download_latest_build(
-        self.out_dir)
+    clusterfuzz_build_dir = self.clusterfuzz_deployment.download_latest_build()
     if not clusterfuzz_build_dir:
       # Crash is reproducible on PR build and we can't test on a recent
       # ClusterFuzz/OSS-Fuzz build.
@@ -278,22 +300,21 @@ class FuzzTarget:
       logging.info('The crash is reproducible on previous build. '
                    'Code change (pr/commit) did not introduce crash.')
       return False
-    logging.info('The crash doesn\'t reproduce on previous build. '
+    logging.info('The crash is not reproducible on previous build. '
                  'Code change (pr/commit) introduced crash.')
     return True
 
-  def get_testcase(self, error_bytes):
-    """Gets the file from a fuzzer run stacktrace.
 
-    Args:
-      error_bytes: The bytes containing the output from the fuzzer.
+def get_testcase(stderr_bytes):
+  """Gets the file from a fuzzer run stacktrace.
 
-    Returns:
-      The path to the testcase or None if not found.
-    """
-    # TODO(metzman): Stop parsing and use libFuzzers' artifact_prefix option
-    # instead.
-    match = re.search(rb'\bTest unit written to \.\/([^\s]+)', error_bytes)
-    if match:
-      return os.path.join(self.out_dir, match.group(1).decode('utf-8'))
-    return None
+  Args:
+    stderr_bytes: The bytes containing the output from the fuzzer.
+
+  Returns:
+    The path to the testcase or None if not found.
+  """
+  match = re.search(rb'\bTest unit written to (.+)', stderr_bytes)
+  if match:
+    return match.group(1).decode('utf-8')
+  return None
